CT-26: Fix all Greptile review issues (5 comments)
1. CRITICAL: require_editor_access! → require_coordinator_or_above!
   Was NoMethodError in production, blocking entire verification workflow

2. Bulk verify audit logs: capture old statuses BEFORE update_all
   Was recording stale data after bulk update

3. Migration: verified_by_user_id integer → bigint
   Match users table PK type to prevent truncation risk

4. Frontend: permission-gate verify buttons behind canEdit + add try/catch
   Buttons now hidden for read-only roles, errors shown on failure

5. Validate supporter_ids param in bulk_verify
   Guard against nil/empty/non-array input

+++ api/db/migrate/20260214042528_add_verification_to_supporters.rb
@@ -1,7 +1,7 @@
 class AddVerificationToSupporters < ActiveRecord::Migration[8.1]
   def change
     add_column :supporters, :verification_status, :string, default: "unverified", null: false
-    add_column :supporters, :verified_by_user_id, :integer
+    add_column :supporters, :verified_by_user_id, :bigint
     add_column :supporters, :verified_at, :datetime
 
     add_index :supporters, :verification_status
